pyproject.toml
[tool.poetry]
name = "bdd-det-yolov5"
version = "0.1.0"
description = "Detection pipeline for BBD dataset"
authors = ["danish <danish.saba@gmail.com>"]
readme = "README.md"
packages = [{include = "bdd_det_yolov5"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
torchvision = "^0.17.0"
pyyaml = "^6.0.1"
tqdm = "^4.66.2"
ultralytics = "^8.1.14"
loguru = "^0.7.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
